requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
chromadb>=0.5.0
mitreattack-python>=2.0.0
requests>=2.31.0
langsmith>=0.1.0
langfuse>=2.0.0
arize-phoenix>=4.0.0
openinference-instrumentation-langchain>=0.1.0
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp-proto-grpc>=1.25.0
opentelemetry-api>=1.25.0
google-generativeai>=0.7.0
python-dotenv>=1.0.0
click>=8.1.0
pydantic>=2.0.0
rich>=13.0.0
